rename authorizeProjectAdmin -> authorizeRegionAdmin
