+++ src/bikelane-verifications/mutations/createBikelaneVerification.ts
@@ -1,6 +1,6 @@
 import { resolver } from '@blitzjs/rpc'
 import db from 'db'
-import { authorizeRegionAdmin } from 'src/authorization/authorizeProjectAdmin'
+import { authorizeRegionAdmin } from 'src/authorization/authorizeRegionAdmin'
 import getRegionIdBySlug from 'src/regions/queries/getRegionIdBySlug'
 import { z } from 'zod'
 import { CreateVerificationSchema, VerificationSchema } from '../schemas'
